requirements.txt
streamlit
requests
python-dotenv
groq
pandas